Rename Table to Tables
Apparently, react styleguidist's "section" implementation has some
weird behavior when you have a component that shares the same name as
its parent directory. In this case, it was rendering the top-level Table
example, but none of the other components were appearing. This isn't a
huge problem here, but I do think it's helpful to have the Prop &
Methods sections for those component, and when we start adding more
components to the library it could be a bigger issue.

As a general rule going forward, we should try to give directories
plural names, if it makes sense, or find another way to differentiate
between the directory and component. A good example: Button/BasicButton
