+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './Button';
 export * from './Layout';
 export * from './Theme';
-export * from './Table';
+export * from './Tables';
 export * from './Layout';
